Added drop_player command.
